pls wokr babe<3

+++ src/main/java/com/sasha/xdolf/events/ClientOverlayRenderEvent.java
@@ -6,13 +6,13 @@ import com.sasha.eventsys.SimpleEvent;
  * Created by Sasha on 09/08/2018 at 11:30 AM
  **/
 public class ClientOverlayRenderEvent extends SimpleEvent {
-    private int partialTicks;
+    private float partialTicks;
 
-    public ClientOverlayRenderEvent(int partialTicks) {
+    public ClientOverlayRenderEvent(float partialTicks) {
         this.partialTicks = partialTicks;
     }
 
-    public int getPartialTicks() {
+    public float getPartialTicks() {
         return partialTicks;
     }
 }
